Change latestChapter type of Comic form DOUBLE to SHORT/SMALLINT

+++ src/Server/BusinessLogicLayer/Models/ComicModel.cs
@@ -5,23 +5,23 @@ namespace MangaManagementAPI.Models;
 
 public class ComicModel
 {
-	public Guid ComicIdentifier { get; set; }
+    public Guid ComicIdentifier { get; set; }
 
-	public string Name { get; set; }
+    public string Name { get; set; }
 
-	public string Description { get; set; }
+    public string Description { get; set; }
 
-	public string Avatar { get; set; }
+    public string Avatar { get; set; }
 
-	public DateOnly PublishDate { get; set; }
+    public DateOnly PublishDate { get; set; }
 
-	public double LatestChapter { get; set; }
+    public short LatestChapter { get; set; }
 
-	public ICollection<ReviewComicModel> ReviewComics { get; set; } = new List<ReviewComicModel>();
+    public ICollection<ReviewComicModel> ReviewComics { get; set; } = new List<ReviewComicModel>();
 
-	public ICollection<ComicSavingModel> ComicSavings { get; set; } = new List<ComicSavingModel>();
+    public ICollection<ComicSavingModel> ComicSavings { get; set; } = new List<ComicSavingModel>();
 
-	public ICollection<ChapterModel> Chapters { get; set; } = new List<ChapterModel>();
+    public ICollection<ChapterModel> Chapters { get; set; } = new List<ChapterModel>();
 
-	public ICollection<ComicCategoryModel> ComicCategories { get; set; } = new List<ComicCategoryModel>();
+    public ICollection<ComicCategoryModel> ComicCategories { get; set; } = new List<ComicCategoryModel>();
 }
